chore: reorganize recap datasets and workers
